+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Add loadEnvrc command which lets the user load any .envrc file
 
 ## [0.14.0] - 2023-06-23
 ### Changed